re_view_time_series
Module series_query

Shared functionality for querying time series data.

Functions

all_scalars_indices

allocate_plot_points

Allocates all points for the series.

collect_colors

Collects colors for the series into pre-allocated plot points.

collect_radius_ui

Collects radius_ui for the series into pre-allocated plot points.

collect_scalars

Allocates scalars per series into pre-allocated plot points.

collect_series_name

Collects series names for the series into pre-allocated plot points.

collect_series_visibility

Queries the visibility flags for all series in a query.

determine_num_series

Determines how many series there are in the scalar chunks.

Type Aliases

PlotPointsPerSeries 🔒